Refactor Receive Components
- Divided Receive into `Receive` and `ReceiveAmount`
- Exported `ShareQR` button as a component
- Updated `Dashboard` with new routes
- Exported `styles` for all the _receive_ screens
